requirements_enhanced.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.3.1
django-jazzmin==3.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-multitenant==2.3.0
django-bootstrap5==23.3
psycopg2-binary==2.9.10
html5lib==1.1
idna==3.10
lxml==5.3.1
mysqlclient==2.2.7
oscrypto==1.3.0
pillow==10.0.0
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
python-bidi==0.6.6
python-dotenv==1.1.0
PyYAML==6.0.2
qrcode==7.3.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.17

# Security enhancements
django-otp==1.2.2
django-two-factor-auth==1.15.3
django-axes==6.1.1

# Real-time features
channels==4.0.0
daphne==4.0.0

# Analytics and reporting
django-import-export==4.3.7
openpyxl==3.1.2

# Inventory forecasting
pandas==2.0.3
scikit-learn==1.3.0

# Caching and performance
django-cacheops==7.0.1